Plant_Disease_detection_GradioAPP/requirements.txt
gradio==4.44.0
torch
torchvision
timm
Pillow
matplotlib
